# docs/wp03/bee-local-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BEE local API contract",
    "version": "0.4.0",
    "summary": "Existing local routes of the Next.js portal boundary and the Spring API (WP03.1).",
...
  },
...
  "paths": {
...
    "/api/model-applications": {
      "servers": [
        {
          "url": "http://127.0.0.1:8090",
          "description": "Spring API. INTERNAL: bound to 127.0.0.1 and called only by Next.js server code with the session's access token."
        }
      ],
      "x-bee-audience": "internal",
      "get": {
        "operationId": "springListModelApplications",
        "tags": [
          "spring-internal"
        ],
        "security": [
          {
            "springBearer": []
          }
        ],
        "summary": "Model applications the caller may read (list only; no filters).",
        "description": "Scope is computed from database rows only: partner roles see their active filing organisation's records; IAME/Reviewer see records with an active assignment at the current stage; Finance, Programme, Director and Secretary see only their own stage. Query parameters are ignored and cannot widen scope.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped list (may be empty).",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApplicationList"
                },
                "example": {
                  "items": [
                    {
                      "id": "00000000-0000-4000-c000-000000000002",
                      "reference": "LOCAL-MA-0002",
                      "organisation": "NOVA",
                      "brandName": "Nova Cool",
                      "category": "RAC",
                      "modelNumber": "NC-RAC-18F",
                      "state": "fee_due",
                      "version": 1,
                      "readBasis": [
                        "own-org"
                      ]
                    }
                  ],
                  "count": 1,
                  "authority": "spring-database"
                }
              }
            }
          },
          "401": {
            "description": "No valid bearer token.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "unauthenticated": {
                    "value": {
                      "error": "unauthenticated",
                      "message": "A valid access token is required."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "unauthenticated"
            ]
          },
          "403": {
            "description": "Denied by the Spring database or MFA policy.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "mfa_required": {
                    "value": {
                      "error": "mfa_required",
                      "message": "Sign-in must include a verified one-time code."
                    }
                  },
                  "no_active_account": {
                    "value": {
                      "error": "no_active_account",
                      "message": "There is no active BEE account for this identity."
                    }
                  },
                  "no_effective_role": {
                    "value": {
                      "error": "no_effective_role",
                      "message": "There is no active BEE role for this identity."
                    }
                  },
                  "no_read_scope": {
                    "value": {
                      "error": "no_read_scope",
                      "message": "This role has no read access to model applications."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "mfa_required",
              "no_active_account",
              "no_effective_role",
              "no_read_scope"
            ]
          },
          "500": {
            "description": "Unexpected failure (no detail).",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "internal_error": {
                    "value": {
                      "error": "internal_error",
                      "message": "The request could not be completed."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "internal_error"
            ]
          },
          "503": {
            "description": "Database unavailable.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "service_unavailable": {
                    "value": {
                      "error": "service_unavailable",
                      "message": "The service is temporarily unavailable. Try again later."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "service_unavailable"
            ]
          }
        }
      },
      "post": {
        "operationId": "springCreateModelApplicationDraft",
        "tags": [
          "spring-internal"
        ],
        "security": [
          {
            "springBearer": []
          }
        ],
        "summary": "Create a draft model application.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApplication"
                }
              }
            }
          },
          "401": {
            "description": "No token.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "unauthenticated": {
                    "value": {
                      "error": "unauthenticated",
                      "message": "A valid access token is required."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "unauthenticated"
            ]
          },
          "403": {
            "description": "Denied.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "mfa_required": {
                    "value": {
                      "error": "mfa_required",
                      "message": "Sign-in must include a verified one-time code."
                    }
                  },
                  "no_active_account": {
                    "value": {
                      "error": "no_active_account",
                      "message": "There is no active BEE account for this identity."
                    }
                  },
                  "no_effective_role": {
                    "value": {
                      "error": "no_effective_role",
                      "message": "There is no active BEE role for this identity."
                    }
                  },
                  "no_write_scope": {
                    "value": {
                      "error": "no_write_scope",
                      "message": "This role cannot create or edit model application drafts."
                    }
                  },
                  "brand_not_permitted": {
                    "value": {
                      "error": "brand_not_permitted",
                      "message": "This brand is not available to your organisation."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "mfa_required",
              "no_active_account",
              "no_effective_role",
              "no_write_scope",
              "brand_not_permitted"
            ]
          },
          "409": {
            "description": "Idempotency conflict.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "idempotency_key_conflict": {
                    "value": {
                      "error": "idempotency_key_conflict",
                      "message": "This Idempotency-Key was already used with a different request body."
                    }
                  },
                  "idempotency_in_progress": {
                    "value": {
                      "error": "idempotency_in_progress",
                      "message": "A request with this Idempotency-Key is still in progress."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "idempotency_key_conflict",
              "idempotency_in_progress"
            ]
          },
          "422": {
            "description": "Invalid input.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "validation_failed": {
                    "value": {
                      "error": "validation_failed",
                      "message": "The request could not be accepted."
                    }
                  },
                  "idempotency_key_required": {
                    "value": {
                      "error": "idempotency_key_required",
                      "message": "An Idempotency-Key header is required for this request."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "validation_failed",
              "idempotency_key_required"
            ]
          },
          "503": {
            "description": "Database unavailable.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "service_unavailable": {
                    "value": {
                      "error": "service_unavailable",
                      "message": "The service is temporarily unavailable. Try again later."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "service_unavailable"
            ]
          }
        }
      }
    },
...
    "/api/runtime/model-applications/{id}": {
      "servers": [
        {
          "url": "http://127.0.0.1:3100",
          "description": "Next.js portal. BROWSER-FACING same-origin routes; the browser never holds a token."
        }
      ],
      "x-bee-audience": "browser",
      "get": {
        "operationId": "runtimeReadModelApplication",
        "tags": [
          "next-browser"
        ],
        "security": [
          {
            "portalSession": []
          }
        ],
        "summary": "One model application for this browser's server session, if Spring lets it read it.",
        "description": "Next.js does not check the ID's shape, cache, or redirect, so it never answers for a record itself. The ID goes to Spring as one path segment; a value that cannot travel safely as one segment is replaced by a fixed non-UUID that Spring answers like any malformed ID. Out-of-scope, unknown, malformed and stale-assignment IDs therefore get the same 404 after Spring's identity and scope checks.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorrelationId"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Model application UUID; any other value is answered as not_found by Spring.",
            "schema": {
              "type": "string"
            },
            "example": "00000000-0000-4000-c000-000000000002"
          }
        ],
        "responses": {
          "200": {
            "description": "Readable record.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelApplication"
                },
                "example": {
                  "id": "00000000-0000-4000-c000-000000000002",
                  "reference": "LOCAL-MA-0002",
                  "organisation": "NOVA",
                  "brandName": "Nova Cool",
                  "category": "RAC",
                  "modelNumber": "NC-RAC-18F",
                  "state": "fee_due",
                  "version": 1,
                  "readBasis": [
                    "own-org"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "No usable session (cookie cleared), or Spring rejected the session's token.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "no_session": {
                    "value": {
                      "error": "no_session",
                      "message": "Sign in to continue."
                    }
                  },
                  "session_expired": {
                    "value": {
                      "error": "session_expired",
                      "message": "Your session has ended. Sign in again."
                    }
                  },
                  "unauthenticated": {
                    "value": {
                      "error": "unauthenticated",
                      "message": "A valid access token is required."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "no_session",
              "session_expired",
              "unauthenticated"
            ]
          },
          "403": {
            "description": "Spring denied the identity or the role has no read rule.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "mfa_required": {
                    "value": {
                      "error": "mfa_required",
                      "message": "Sign-in must include a verified one-time code."
                    }
                  },
                  "no_active_account": {
                    "value": {
                      "error": "no_active_account",
                      "message": "There is no active BEE account for this identity."
                    }
                  },
                  "no_effective_role": {
                    "value": {
                      "error": "no_effective_role",
                      "message": "There is no active BEE role for this identity."
                    }
                  },
                  "no_read_scope": {
                    "value": {
                      "error": "no_read_scope",
                      "message": "This role has no read access to model applications."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "mfa_required",
              "no_active_account",
              "no_effective_role",
              "no_read_scope"
            ]
          },
          "404": {
            "description": "Not available to this session. Unknown, malformed, out-of-scope and stale-assignment IDs are indistinguishable.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "error": "not_found",
                      "message": "No such record is available to you."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "not_found"
            ]
          },
          "405": {
            "description": "Method not supported on this route.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "method_not_allowed": {
                    "value": {
                      "error": "method_not_allowed",
                      "message": "This method is not supported for this resource."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "method_not_allowed"
            ]
          },
          "502": {
            "description": "Spring answered outside the contract. Nothing from Spring's body is forwarded.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "invalid_api_response": {
                    "value": {
                      "error": "invalid_api_response",
                      "message": "The BEE service returned an unexpected response."
                    }
                  },
                  "api_error": {
                    "value": {
                      "error": "api_error",
                      "message": "The BEE service could not complete the request."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "invalid_api_response",
              "api_error"
            ]
          },
          "503": {
            "description": "A dependency is unavailable; retry later. Session kept.",
            "headers": {
              "X-Correlation-Id": {
                "$ref": "#/components/headers/CorrelationId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "api_unreachable": {
                    "value": {
                      "error": "api_unreachable",
                      "message": "The BEE service is not reachable. Try again later."
                    }
                  },
                  "identity_unavailable": {
                    "value": {
                      "error": "identity_unavailable",
                      "message": "The identity service is not reachable. Try again later."
                    }
                  },
                  "service_unavailable": {
                    "value": {
                      "error": "service_unavailable",
                      "message": "The service is temporarily unavailable. Try again later."
                    }
                  }
                }
              }
            },
            "x-error-codes": [
              "api_unreachable",
              "identity_unavailable",
              "service_unavailable"
            ]
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "springBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak bee-local access token. Held only by the Next.js server session; never sent to or by the browser."
      },
      "portalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bee_session",
        "description": "Opaque random session ID (HttpOnly, SameSite=Lax). Created only after password + TOTP sign-in and a Spring 200 on /api/me."
      }
    },
    "parameters": {
      "CorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "description": "Optional caller ID matching ^[A-Za-z0-9-]{1,64}$. Kept if safe; otherwise replaced by a new UUID. Never an authorization input.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9-]{1,64}$"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9-]{16,64}$"
        }
      }
    },
    "headers": {
      "CorrelationId": {
        "required": true,
        "description": "Present on every response, success or denial. Next.js accepts a safe incoming value or issues a UUID, sends it to Spring, and returns the same value. Spring applies the same rule to direct callers. Logged with method, path (no query), status and duration only.",
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9-]{1,64}$"
        }
      },
      "NoStore": {
        "required": true,
        "description": "Contains no-store.",
        "schema": {
          "type": "string",
          "pattern": "no-store"
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "message"
        ],
        "description": "The one error body of both layers. The correlation ID is in the X-Correlation-Id header, not the body. message is fixed per code and never includes record IDs, tokens, SQL, stack traces or Keycloak detail.",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "api_error",
              "api_unreachable",
              "brand_not_permitted",
              "cross_origin",
              "denied_by_default",
              "idempotency_in_progress",
              "idempotency_key_conflict",
              "idempotency_key_required",
              "identity_unavailable",
              "internal_error",
              "invalid_api_response",
              "method_not_allowed",
              "mfa_required",
              "no_active_account",
              "no_effective_role",
              "no_read_scope",
              "no_session",
              "no_write_scope",
              "not_editable",
              "not_found",
              "service_unavailable",
              "session_expired",
              "unauthenticated",
              "validation_failed",
              "version_conflict"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ModelApplication": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "reference",
          "organisation",
          "brandName",
          "category",
          "modelNumber",
          "state",
          "version",
          "readBasis"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string"
          },
          "organisation": {
            "type": "string"
          },
          "brandName": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "modelNumber": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "draft",
              "fee_due",
              "iame_scrutiny",
              "bee_scrutiny",
              "rating",
              "director_review",
              "secretary_approval",
              "approved",
              "returned",
              "rejected"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 0,
            "description": "Persisted row version; the expected-version input of future writes (409 version_conflict)."
          },
          "readBasis": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "pattern": "^(own-org|assigned|stage:[a-z_]+)$"
            },
            "description": "Which database rule allowed the read (own-org, assigned, stage:<state>)."
          },
          "brandId": {
            "type": "string",
            "format": "uuid"
          },
          "principalOrganisation": {
            "type": "string"
          }
        }
      },
      "ModelApplicationList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "count",
          "authority"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelApplication"
            }
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "authority": {
            "type": "string",
            "const": "spring-database"
          }
        }
      },
...
      "ModelApplicationCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "brandId",
          "category",
          "modelNumber"
        ],
        "properties": {
          "brandId": {
            "type": "string",
            "format": "uuid"
          },
          "category": {
            "const": "RAC"
          },
          "modelNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        }
      },
...
    }
  },
...
}